Time: 16 ms (29.21%), Space: 36.7 MB (67.39%) - LeetHub

+++ 1976-number-of-ways-to-arrive-at-destination/1976-number-of-ways-to-arrive-at-destination.cpp
@@ -19,7 +19,7 @@ public:
 
         while(!pq.empty()) {
             int node = pq.top().second;
-            int time = pq.top().first;
+            long long time = pq.top().first;
             vector<pair<int, int>> neighbors = graph[node];
             pq.pop();
